refactor(UnifiedCard): extract SimpleCard and AdvancedCard sub-components (#135)
